1_django-movies/requirements.txt
asgiref==3.6.0
distlib==0.3.6
Django==4.2
djangorestframework==3.14.0
filelock==3.11.0
mysql-connector-python==8.0.31
platformdirs==3.2.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
virtualenv==20.21.0
